+++ Solution/RiskOfChaos/Components/ICustomPingBehavior.cs
@@ -1,0 +1,11 @@
+﻿using RoR2.UI;
+
+namespace RiskOfChaos.Components
+{
+    public interface ICustomPingBehavior
+    {
+        void OnPingAdded(PingIndicator pingIndicator);
+
+        void OnPingRemoved(PingIndicator pingIndicator);
+    }
+}