+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+**-.-.- Changes:**
+
+* Spawn Fake Teleporter:
+  * Fixed telporter particles not scaling when "Scale Teleporter Particles" setting is enabled
+
 **2.4.1 Changes:**
 
 * Added 1 effect: