+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Spawn Fake Teleporter:
   * Fixed telporter particles not scaling when "Scale Teleporter Particles" setting is enabled
 
+* Bouncy Projectiles:
+  * Fixed bullets sometimes dealing damage several times to the same target
+
 **2.4.1 Changes:**
 
 * Added 1 effect: